Add example sim object from Sean Johnson
First extension is flux in nJy. Second extension is wavelength in nm.
Other extensions can be ignored.

This is included for the sake of testing. A full collection of
simulated spectra will be added at a later point, and it will
probably be more than we should store in this git-lfs repo.